[update] Station15

+++ app/Http/Controllers/AdminController.php
@@ -119,6 +119,9 @@ class AdminController extends Controller
     public function edit($id)
     {
         $movie = Movie::with('genre')->find($id);
+        if ($movie == null) {
+            return response('', 404);
+        }
         return view('adminEditMovie', compact('movie'));
     }
 
@@ -143,48 +146,111 @@ class AdminController extends Controller
 
     public function getallschedules()
     {
-        $movies = Movie::all()->schedule();
+        $movies = Movie::with('schedules')->get();
         return view('adminShowAllSchedule', compact('movies'));
     }
 
     public function getschedulebyid($id)
     {
         $schedule = Schedule::find($id);
+        if ($schedule == null) {
+            return response('', 404);
+        }
         return view('adminShowSchedulebyid', compact('schedule'));
     }
 
-    public function createschedulemenu()
+    public function createschedulemenu($id)
     {
-        return view('adminCreateSchedule');
+        return view('adminCreateSchedule', ['id' => $id]);
     }
 
-    public function storeschedule(Request $request)
+
+    public function createschedule(Request $request, $id)
     {
-        $validated = $request->validate([
-            'movie_id' => 'required',
-            'start_time_date' => 'required',
-            'start_time_time' => 'required',
-            'end_time_date' => 'required',
-            'end_time_time' => 'required',
-        ]);
+        $result = DB::transaction(function () use ($request) {
+            $validated = $request->validate([
+                'movie_id' => 'required',
+                'start_time_date' => ['required', 'date_format:Y-m-d'],
+                'start_time_time' => ['required', 'date_format:H:i'],
+                'end_time_date' => ['required', 'date_format:Y-m-d'],
+                'end_time_time' => ['required', 'date_format:H:i'],
+            ]);
 
-        $schedule = new Schedule;
+            if ($movie == null) {
+                return response('', 404);
+            }
 
-        //carbon check rule
+            $schedule = new Schedule;
+            // $start_time = Carbon::createFromFormat('Y-m-d H:i:s', $request->start_time_date . ' ' . $request->start_time_time);
+//            $start_time = Carbon::create($request->start_time_date . ' ' . $request->start_time_time);
+            $start_time = $request->start_time_date . ' ' . $request->start_time_time;
+            // $end_time = Carbon::createFromFormat('Y-m-d H:i:s', $request->end_time_date . ' ' . $request->end_time_time);
+//            $end_time = Carbon::create($request->end_time_date . ' ' . $request->end_time_time);
+            $end_time = $request->end_time_date . ' ' . $request->end_time_time;
 
-        $start_time = Carbon::parse($request->start_time_date . ' ' . $request->start_time_time);
-        $end_time = Carbon::parse($request->end_time_date . ' ' . $request->end_time_time);
 
-        $schedule->movie_id = $request->movie_id;
-        $schedule->start_time = $start_time;
-        $schedule->end_time = $end_time;
+            $schedule->movie_id = $request->movie_id;
+            $schedule->start_time = $start_time;
+            $schedule->end_time = $end_time;
 
-        $schedule->save();
+            $schedule->save();
+        });
+        if ($result == 0) {
+            $flashmessage = "エラーが発生しました.";
+//             return redirect()->route('schedule.create', $id)->with('message.success', $flashmessage);
+            return redirect('error', 500);
+        } else {
+            $flashmessage = "スケジュールを作成しました.";
+            return redirect()->route('schedule.create', $id)->with('message.success', $flashmessage);
+        }
     }
+
+    public function storeschedule(Request $request, $id)
+    {
+        $movie = Movie::find($id);
+        if ($movie == null) {
+            return response('', 404);
+        }
+        $result = DB::transaction(function () use ($request, $id) {
+            $validated = $request->validate([
+                'movie_id' => 'required',
+                'start_time_date' => ['required', 'date_format:Y-m-d'],
+                'start_time_time' => ['required', 'date_format:H:i'],
+                'end_time_date' => ['required', 'date_format:Y-m-d'],
+                'end_time_time' => ['required', 'date_format:H:i'],
+            ]);
+
+            $schedule = new Schedule;
+//            $start_time = Carbon::createFromFormat('Y-m-d H:i', $request->start_time_date . ' ' . $request->start_time_time)->format('Y-m-d H:i:s');
+//            $end_time = Carbon::createFromFormat('Y-m-d H:i', $request->end_time_date . ' ' . $request->end_time_time)->format('Y-m-d H:i:s');
+            $start_time = Carbon::parse($request->start_time_date . ' ' . $request->start_time_time)->format('Y-m-d H:i');
+            $end_time = Carbon::parse($request->end_time_date . ' ' . $request->end_time_time)->format('Y-m-d H:i');
+
+
+            $schedule->movie_id = $id;
+            $schedule->start_time = $start_time;
+            $schedule->end_time = $end_time;
+
+            $schedule->save();
+        });
+        if ($result == 0) {
+            $flashmessage = "エラーが発生しました.";
+            return redirect()->route('schedule.create', $id)->with('message.success', $flashmessage);
+        } else {
+            $flashmessage = "スケジュールを作成しました.";
+            return redirect()->route('schedule.create', $id)->with('message.success', $flashmessage);
+        }
+    }
+
 
     public function editschedule($scheduleId)
     {
         $schedule = Schedule::find($scheduleId);
+
+        if ($schedule == null) {
+            return response('', 404);
+        }
+
         return view('adminEditSchedule', compact('schedule'));
     }
 
@@ -192,39 +258,55 @@ class AdminController extends Controller
     {
         $schedule = Schedule::find($id);
 
+        if ($schedule == null) {
+            return response('', 404);
+        }
+
         $validated = $request->validate([
-            'start_time_date' => 'required',
-            'start_time_time' => 'required',
-            'end_time_date' => 'required',
-            'end_time_time' => 'required',
+            'movie_id' => 'required',
+            'start_time_date' => ['required', 'date_format:Y-m-d'],
+            'start_time_time' => ['required', 'date_format:H:i'],
+            'end_time_date' => ['required', 'date_format:Y-m-d'],
+            'end_time_time' => ['required', 'date_format:H:i'],
         ]);
 
         //check format is correct(YYYY-MM-DD H:i)
         $rules =
             //日付時刻を結合
-        $start_time = Carbon::parse($request->start_time_date . ' ' . $request->start_time_time);
-        $end_time = Carbon::parse($request->end_time_date . ' ' . $request->end_time_time);
+        $start_time = Carbon::parse($request->start_time_date . ' ' . $request->start_time_time)->format('Y-m-d H:i');
+        $end_time = Carbon::parse($request->end_time_date . ' ' . $request->end_time_time)->format('Y-m-d H:i');
 
         $schedule->start_time = $start_time;
         $schedule->end_time = $end_time;
 
         //更新処理
         $schedule->save();
+        if ($schedule == null) {
+            return redirect()->route('schedule.edit', $id)->with('message.success', '更新に失敗しました.');
+        } else {
+            return redirect()->route('schedule.edit', $id)->with('message.success', '更新に成功しました.');
+        }
     }
 
 
-    public function destroyallschedule($id)
+    public function destroyschedule($id)
     {
         $schedule = Schedule::find($id);
+        if ($schedule == null) {
+            return response('', 404);
+        }
         $schedule->delete();
-
+        return redirect()->route('schedule.getall')->with('message.success', '削除に成功しました.');
     }
 
     public function adminmovieinfo($id)
     {
-        $movie = Movie::find($id);
-        $schedules = $movie->schedules()->orderBy('start_time')->get();
-        return view('adminMovieInfo', compact('movie', 'schedules'));
+//        $movie = Movie::find($id);
+        $movie = Movie::with('schedules')->find($id);
+        if ($movie == null) {
+            return response('', 404);
+        }
+//        $schedules = $movie->schedules()->orderBy('start_time')->get();
+        return view('adminMovieInfo', compact('movie'));
     }
-
 }
